Create Three ball monty
This is a classic three ball monty game where you have to guess the position of the ball. Run the program and enter your guess of the index position of the ball. If your guess is right, the program will say you won or if you guessed wrong the program will say you lost and gives you the list of correct position of there ball